+++ src/app/designsystem/page.tsx
@@ -6,33 +6,19 @@ import {
   DefaultBanner,
   Toolbar,
   ButtonList,
-  DesignSystemCardContainer,
   Footer,
-  DesignSystemCard,
-  BadgeLabel,
-  Button,
   EmptyState,
 } from "@/components";
-import { BadgeLabelFeedback } from "@/components/Badge/Badge.types";
-import { ButtonStyle } from "@/components/Button/Button.types";
 import { BANNER_TEXT, DESIGNSYSTEM_PAGE_TEXT } from "@/constants/messages";
 import { useTokenStore } from "@/hooks/store/useTokenStore";
 import { useDesignSystemInfiniteQuery } from "@/hooks/api/designSystem/useDesignSystemInfiniteQuery";
-import { IDesignSystemData } from "@/types/api/designSystem";
 import { useRef } from "react";
 import { useObserver } from "@/hooks/api/common/useObserver";
-import { DESIGN_SYSTEM_CHIP_GROUP } from "@/constants/chipGroup";
 import {
-  ContentFilter,
-  DesignSystemFilterType,
-  DeviceFilter,
-  TechFilter,
-} from "@/types/enum/designSystemFilters";
-import {
-  contentFilterLabels,
-  deviceFilterLabels,
-  techFilterLabels,
-} from "@/constants/designSystemFilterLabel";
+  DesignSystemCardContainer,
+  DesignSystemCard,
+} from "@/components/Pages";
+import { IDesignSystemData } from "@/types/api/designSystem";
 
 export default function DesignSystem() {
   const { accessToken } = useTokenStore();
@@ -69,73 +55,12 @@ export default function DesignSystem() {
       </Toolbar>
       <DesignSystemCardContainer>
         {designSystemList?.pages.map((page) =>
-          page.content.map((designSystem: IDesignSystemData) => {
-            const deviceLabels = designSystem.filters.filter(
-              (value) => value.type === DesignSystemFilterType.DEVICE
-            );
-            const labels = designSystem.filters.filter(
-              (value) =>
-                value.type === DesignSystemFilterType.TECH ||
-                value.type === DesignSystemFilterType.CONTENT
-            );
-            const platformLabels = designSystem.filters.filter(
-              (value) => value.type === DesignSystemFilterType.PLATFORM
-            );
-
-            return (
-              <DesignSystemCard
-                key={`designSystemcard-${designSystem.name}`}
-                designSystemName={designSystem.name}
-                organizationName={designSystem.organizationName}
-                descriptionText={designSystem.description}
-                $bookmarkCount="999+"
-                deviceLabels={deviceLabels.map((deviceLabel) =>
-                  deviceLabel.values.map((deviceName) => (
-                    <BadgeLabel
-                      key={`deviceLabel-${deviceName}`}
-                      $variant="labelOnly"
-                      text={deviceFilterLabels[deviceName as DeviceFilter]}
-                      $feedback={BadgeLabelFeedback.NONE}
-                      $style="solid"
-                      $size="xs"
-                    />
-                  ))
-                )}
-                labels={labels.map((label) =>
-                  label.values.map((labelName) => (
-                    <BadgeLabel
-                      key={`label-${label}`}
-                      $variant="labelOnly"
-                      text={
-                        label.type === "TECH"
-                          ? techFilterLabels[labelName as TechFilter]
-                          : contentFilterLabels[labelName as ContentFilter]
-                      }
-                      $feedback={BadgeLabelFeedback.NONE}
-                      $style="transparent"
-                      $size="xs"
-                    />
-                  ))
-                )}
-                platformButtons={platformLabels.map((platformLabel) =>
-                  platformLabel.values.map((platformName) => (
-                    <Button
-                      key={`platformButton-${platformName}`}
-                      text={platformName}
-                      $size="md"
-                      $buttonType="iconButton"
-                      $leftIcon={
-                        DESIGN_SYSTEM_CHIP_GROUP.platform.contents.find(
-                          (content) => content.responseName === platformName
-                        )?.icon
-                      }
-                      $buttonStyle={ButtonStyle.OutlinedSecondary}
-                    />
-                  ))
-                )}
-              />
-            );
-          })
+          page.content.map((designSystemData: IDesignSystemData) => (
+            <DesignSystemCard
+              key={designSystemData.name}
+              designSystem={designSystemData}
+            />
+          ))
         )}
       </DesignSystemCardContainer>
       {isLoading && <EmptyState text={DESIGNSYSTEM_PAGE_TEXT.loading} />}
