+++ src/components/index.ts
@@ -29,12 +29,11 @@ export { default as Toolbar } from "./Toolbar/Toolbar";
 export { ButtonList } from "./Toolbar/Toolbar.variant";
 export { ChipList } from "./Toolbar/Toolbar.variant";
 export { default as ComponentCard } from "./Card/Card.component";
-export { default as DesignSystemCard } from "./Card/Card.designSystem";
+export { default as CardDesignSystem } from "./Card/Card.designSystem";
 export { default as Dialog } from "./Dialog/Dialog";
 export { default as Layout } from "./Layout/Layout";
 export { default as CardContainer } from "./Pages/Component/CardContainer";
 export { default as ImageContainer } from "./Pages/Onboarding/ImageContainer";
-export { default as DesignSystemCardContainer } from "./Pages/DesignSystem/DesignCardContainer";
 export { default as EmptyState } from "./EmptyState/EmptyState";
 export { default as Pagination } from "./Pagination/Pagination";
 export { default as Accordion } from "./Accordion/Accordion";
